third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/container-size-invalidation-after-load-expected.txt

This is a testharness.js-based test.
FAIL @container: invalidation of container size after load event assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
